+++ js/apps/admin-ui/src/identity-providers/add/DetailSettings.tsx
@@ -58,6 +58,7 @@ import { OIDCAuthentication } from "./OIDCAuthentication";
 import { OIDCGeneralSettings } from "./OIDCGeneralSettings";
 import { ReqAuthnConstraints } from "./ReqAuthnConstraintsSettings";
 import { SamlGeneralSettings } from "./SamlGeneralSettings";
+import { OpenIdFederationSettings } from "./OpenIdFederationSettings";
 
 type HeaderProps = {
   onChange: (value: boolean) => void;
@@ -340,7 +341,9 @@ export default function DetailSettings() {
     return <KeycloakSpinner />;
   }
 
-  const isOIDC = provider.providerId!.includes("oidc");
+  const isOIDC =
+    provider.providerId!.includes("oidc") ||
+    provider.providerId!.includes("openid-federation");
   const isSAML = provider.providerId!.includes("saml");
 
   const loader = async () => {
@@ -392,6 +395,11 @@ export default function DetailSettings() {
           {isSAML && <SamlGeneralSettings id={alias} isAliasReadonly />}
         </FormAccess>
       ),
+    },
+    {
+      title: t("openIdFederationSettings"),
+      isHidden: !provider.providerId!.includes("openid-federation"),
+      panel: <OpenIdFederationSettings readOnly={true} />,
     },
     {
       title: t("oidcSettings"),