+++ js/apps/admin-ui/src/identity-providers/component/RedirectUrl.tsx
@@ -5,10 +5,17 @@ import { HelpItem } from "ui-shared";
 import { adminClient } from "../../admin-client";
 import { useRealm } from "../../context/realm-context/RealmContext";
 import { addTrailingSlash } from "../../util";
+import { useFormContext, useWatch } from "react-hook-form";
+import IdentityProviderRepresentation from "libs/keycloak-admin-client/lib/defs/identityProviderRepresentation";
 
 export const RedirectUrl = ({ id }: { id: string }) => {
   const { t } = useTranslation("identity-providers");
+  const { control } = useFormContext<IdentityProviderRepresentation>();
 
+  const providerId = useWatch({
+    control,
+    name: "providerId",
+  });
   const { realm } = useRealm();
   const callbackUrl = `${addTrailingSlash(
     adminClient.baseUrl,
@@ -25,9 +32,12 @@ export const RedirectUrl = ({ id }: { id: string }) => {
       }
       fieldId="kc-redirect-uri"
     >
-      <ClipboardCopy
-        isReadOnly
-      >{`${callbackUrl}/${id}/endpoint`}</ClipboardCopy>
+      <ClipboardCopy isReadOnly>
+        {`${callbackUrl}/` +
+          (providerId === "openid-federation"
+            ? "federation-endpoint"
+            : `${id}/endpoint`)}
+      </ClipboardCopy>
     </FormGroup>
   );
 };