+++ CHANGELOG.md
@@ -15,9 +15,19 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 ### Added
 - OpenId Federation client access policies (UI)
 - OpenID Federation RP with explicit registration (experimental feature) 
+- Show expiration time for OpenId Federation Clients
+- Hide Enable/Disable Toggle for OpenId Federation Clients
+- Show Oidc Details for OpenId Federation Identity Providers
+- Show OpenId Federation Settings for OpenId Federation Identity Providers
+
+### Updated
+- Entity type supported values for OpenId Federation
+- Updated Redirect Uri for OpenId Federation Identity Providers
 
 ### Fixed
 - Return trust anchor and authority_hints in explicit client registration response
+- OpenId Federation Endpoint
+- Hide Create option for OpenId Federation Identity Providers
 
 ## [22.0.13-1.19] - 2025-07-10